refactor(product): split sidebar hosts back under the structure thresholds
The pinning feature pushed MainSidebar past its allowlisted 430 lines and
WorkspaceItem past 400. Split along the documented seams instead of widening
the allowlist: the archive/unarchive confirmation flow moves into a
use-workspace-sidebar-archive-actions workflow hook, the right-click delete
confirmation pane into WorkspaceDeleteConfirmMenu, and the Pin glyph into the
workspace icon module. MainSidebar lands at 364 lines, so its now-stale
frontend-structure allowlist entry ("extraction pending") is removed as the
ratchet requires.
