+++ apps/packages/product-client/src/components/workspace/shell/sidebar/MainSidebar.tsx
@@ -1,4 +1,4 @@
-import { memo, useCallback, useMemo, useState } from "react";
+import { memo, useCallback, useMemo } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import { useShallow } from "zustand/react/shallow";
 import { useRemoveCloudRepoEnvironment, useRepositories } from "@proliferate/cloud-sdk-react";
@@ -39,7 +39,7 @@ import { useSessionSelectionStore } from "#product/stores/sessions/session-selec
 import { useWorkspaceUiStore } from "#product/stores/preferences/workspace-ui-store";
 import { useWorkspaceDisplayNameActions } from "#product/hooks/workspaces/workflows/use-workspace-display-name-actions";
 import { useWorkspaceSidebarActions } from "#product/hooks/workspaces/workflows/use-workspace-sidebar-actions";
-import { useCloudWorkspaceActions } from "#product/hooks/cloud/workflows/use-cloud-workspace-actions";
+import { useWorkspaceSidebarArchiveActions } from "#product/hooks/workspaces/workflows/use-workspace-sidebar-archive-actions";
 import { useSidebarRepoGroupState } from "#product/hooks/workspaces/facade/use-sidebar-repo-group-state";
 import { useWorkspaceSidebarState } from "#product/hooks/workspaces/derived/use-workspace-sidebar-state";
 import { useSessionActivityReconciler } from "#product/hooks/sessions/lifecycle/use-session-activity-reconciler";
@@ -47,20 +47,12 @@ import {
   buildCloudRepoSettingsHref,
   buildSettingsHref,
 } from "#product/lib/domain/settings/navigation";
-import { cloudWorkspaceSyntheticId } from "#product/lib/domain/workspaces/cloud/cloud-ids";
 import { getShortcutDisplayLabel } from "#product/lib/domain/shortcuts/matching";
 import { buildShortcutRangeLabelById } from "#product/lib/domain/shortcuts/presentation";
 import { startMeasurementOperation } from "#product/lib/infra/measurement/measurement-port";
 import { useShortcutRevealVisible } from "#product/providers/ShortcutRevealProvider";
-import { useToastStore } from "#product/stores/toast/toast-store";
 import { useReleaseNotice } from "#product/hooks/updates/facade/use-release-notice";
 import { useRepositoryHeaderNewChat } from "#product/hooks/workspaces/ui/use-repository-header-new-chat";
-
-interface ArchiveConfirmationState {
-  workspaceId: string;
-  cloudWorkspaceId: string | null;
-  name: string;
-}
 
 export const MainSidebar = memo(function MainSidebar({ showRightBorder = true }: { showRightBorder?: boolean }) {
   useDebugRenderCount("workspace-sidebar");
@@ -81,7 +73,6 @@ export const MainSidebar = memo(function MainSidebar({ showRightBorder = true }:
     data: repoConfigs,
     isPending: isRepoConfigsPending,
   } = useRepositories(cloudActive);
-  const showToast = useToastStore((state) => state.show);
   const removeCloudRepoEnvironment = useRemoveCloudRepoEnvironment();
   const pendingWorkspaceEntry = useSessionSelectionStore((state) => state.pendingWorkspaceEntry);
   const {
@@ -111,19 +102,23 @@ export const MainSidebar = memo(function MainSidebar({ showRightBorder = true }:
   });
   const navigate = useNavigate();
   const location = useLocation();
-  const [archiveConfirmation, setArchiveConfirmation] =
-    useState<ArchiveConfirmationState | null>(null);
   const {
-    archiveCloudWorkspace: archiveCloudWorkspaceRequest,
-    restoreCloudWorkspace: restoreCloudWorkspaceRequest,
-  } = useCloudWorkspaceActions();
+    archiveConfirmation,
+    closeArchiveConfirmation,
+    confirmArchiveWorkspace,
+    handleArchiveWorkspace,
+    handleUnarchiveWorkspace,
+  } = useWorkspaceSidebarArchiveActions({
+    groups,
+    selectedWorkspaceId,
+    selectedLogicalWorkspaceId,
+    onLeaveWorkspace: actions.handleGoHome,
+  });
 
   const isOnWorkflows = location.pathname.startsWith(APP_ROUTES.workflows);
   const isOnWorkspaces = location.pathname === APP_ROUTES.workspaces;
   const isOnHome = location.pathname === APP_ROUTES.home;
-  const archiveWorkspace = useWorkspaceUiStore((s) => s.archiveWorkspace);
   const hideRepoRoot = useWorkspaceUiStore((s) => s.hideRepoRoot);
-  const unarchiveWorkspace = useWorkspaceUiStore((s) => s.unarchiveWorkspace);
   const pinWorkspace = useWorkspaceUiStore((s) => s.pinWorkspace);
   const unpinWorkspace = useWorkspaceUiStore((s) => s.unpinWorkspace);
   const { updateWorkspaceDisplayName } = useWorkspaceDisplayNameActions();
@@ -181,86 +176,6 @@ export const MainSidebar = memo(function MainSidebar({ showRightBorder = true }:
     groups,
     hideRepoRoot,
     removeCloudRepoEnvironment,
-  ]);
-
-  const resolveArchiveTargetForSidebarItem = useCallback((
-    workspaceId: string,
-  ): ArchiveConfirmationState => {
-    for (const group of groups) {
-      const item = group.items.find((candidate) => candidate.id === workspaceId);
-      if (item) {
-        return {
-          workspaceId,
-          cloudWorkspaceId: item.cloudWorkspaceId,
-          name: item.name,
-        };
-      }
-    }
-    return {
-      workspaceId,
-      cloudWorkspaceId: null,
-      name: "this workspace",
-    };
-  }, [groups]);
-
-  const handleArchiveWorkspace = useCallback((workspaceId: string) => {
-    setArchiveConfirmation(resolveArchiveTargetForSidebarItem(workspaceId));
-  }, [resolveArchiveTargetForSidebarItem]);
-
-  const confirmArchiveWorkspace = useCallback(() => {
-    const target = archiveConfirmation;
-    if (!target) {
-      return;
-    }
-    setArchiveConfirmation(null);
-    const shouldLeaveWorkspace = selectedLogicalWorkspaceId === target.workspaceId
-      || selectedWorkspaceId === target.workspaceId
-      || (
-        target.cloudWorkspaceId
-        ? selectedWorkspaceId === cloudWorkspaceSyntheticId(target.cloudWorkspaceId)
-        : false
-      );
-    const cloudWorkspaceId = target.cloudWorkspaceId;
-    if (!cloudWorkspaceId) {
-      archiveWorkspace(target.workspaceId);
-      if (shouldLeaveWorkspace) {
-        actions.handleGoHome();
-      }
-      return;
-    }
-    if (shouldLeaveWorkspace) {
-      actions.handleGoHome();
-    }
-    void archiveCloudWorkspaceRequest(cloudWorkspaceId)
-      .catch((error) => {
-        const message = error instanceof Error ? error.message : "Failed to archive workspace.";
-        showToast(message);
-      });
-  }, [
-    actions,
-    archiveConfirmation,
-    archiveWorkspace,
-    archiveCloudWorkspaceRequest,
-    selectedLogicalWorkspaceId,
-    selectedWorkspaceId,
-    showToast,
-  ]);
-
-  const handleUnarchiveWorkspace = useCallback((workspaceId: string) => {
-    const cloudWorkspaceId = resolveArchiveTargetForSidebarItem(workspaceId).cloudWorkspaceId;
-    if (!cloudWorkspaceId) {
-      unarchiveWorkspace(workspaceId);
-      return;
-    }
-    void restoreCloudWorkspaceRequest(cloudWorkspaceId).catch((error) => {
-      const message = error instanceof Error ? error.message : "Failed to restore workspace.";
-      showToast(message);
-    });
-  }, [
-    resolveArchiveTargetForSidebarItem,
-    restoreCloudWorkspaceRequest,
-    showToast,
-    unarchiveWorkspace,
   ]);
 
   const handleOpenRepoSettings = useCallback((sourceRoot: string) => {
@@ -440,7 +355,7 @@ export const MainSidebar = memo(function MainSidebar({ showRightBorder = true }:
           title="Archive workspace?"
           description={`Move ${archiveConfirmation?.name ?? "this workspace"} out of the main sidebar. It will remain available in Settings -> Archived chats, and safe worktree cleanup may run in the background.`}
           confirmLabel="Archive"
-          onClose={() => setArchiveConfirmation(null)}
+          onClose={closeArchiveConfirmation}
           onConfirm={confirmArchiveWorkspace}
         />
       </ProductSidebarFrame>
